server/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
firebase-admin
python-multipart
python-dotenv
